pyproject.toml
[build-system]
requires = ["setuptools>=77.0.0", "setuptools-scm>=8.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "conjunto"
dynamic = ["version"]
description = "Django application framework/helpers using Tetra, tabler.io, tables2, crispy-forms & more."
authors = [
    { name = "Christian González", email = "christian.gonzalez@nerdocs.at" }
]
license = {file = "LICENSE"}
readme = "README.md"
keywords = ["python", "django", "tetra", "framework", "i18n"]
classifiers = [
  "Development Status :: 2 - Pre-Alpha",
  "Programming Language :: Python"
]
dependencies = [
    "django>=5.1",
    "django-countries",
    "django-phonenumber-field[phonenumberslite]",
    "django-crispy-forms",
    "crispy-bootstrap5",
    "django-tables2",
    "django-versionfield",
    "django-markdownify",
    "django-web-components",
    "django-statici18n",
    "gdaps>=0.10.0",
    "django-extensions>=3.2",
    "Pillow",
    "tetra>=0.2"
]



[project.urls]
Homepage = "https://github.com/nerdocs/conjunto"
#Documentation = "https://conjunto.readthedocs.io"
Repository = "https://github.com/nerdocs/conjunto"


[project.optional-dependencies]
dev = [
    "conjunto[prod]",
    "build",
    "twine",
    "pytest",
    "pytest-mock",
    "pytest-django",
    "pre-commit",
    "black",
]
doc = [
    "mkdocs",
    "mkdocstrings-python",
    "mkdocs-literate-nav",
    "mkdocs-section-index",
    "pymdown-extensions",
    "pygments",
    "mkdocs-api-autonav",
]
prod = [
]

[tool.setuptools.dynamic]
version = {attr = "conjunto.__version__"}